chore(nexus): reorganize filestream into a package (#6041)
