+++ nexus/pkg/filestream/loop_feedback.go
@@ -1,0 +1,10 @@
+package filestream
+
+func (fs *FileStream) addFeedback(reply map[string]interface{}) {
+	fs.feedbackChan <- reply
+}
+
+func (fs *FileStream) loopFeedback(inChan <-chan map[string]interface{}) {
+	for range inChan {
+	}
+}
