+++ nexus/pkg/server/sender.go
@@ -12,6 +12,7 @@ import (
 	"github.com/wandb/wandb/nexus/internal/nexuslib"
 	"github.com/wandb/wandb/nexus/internal/uploader"
 	"github.com/wandb/wandb/nexus/pkg/artifacts"
+	fs "github.com/wandb/wandb/nexus/pkg/filestream"
 	"github.com/wandb/wandb/nexus/pkg/observability"
 
 	"github.com/Khan/genqlient/graphql"
@@ -26,7 +27,7 @@ const (
 )
 
 type ResumeState struct {
-	FileStreamOffset map[chunkFile]int
+	FileStreamOffset fs.FileStreamOffsetMap
 	Error            service.ErrorInfo
 }
 
@@ -52,7 +53,7 @@ type Sender struct {
 	graphqlClient graphql.Client
 
 	// fileStream is the file stream
-	fileStream *FileStream
+	fileStream *fs.FileStream
 
 	// uploader is the file uploader
 	uploader *uploader.Uploader
@@ -175,12 +176,13 @@ func (s *Sender) sendRunStart(_ *service.RunStartRequest) {
 	if !s.settings.GetXOffline().GetValue() {
 		fsPath := fmt.Sprintf("%s/files/%s/%s/%s/file_stream",
 			s.settings.GetBaseUrl().GetValue(), s.RunRecord.Entity, s.RunRecord.Project, s.RunRecord.RunId)
-		s.fileStream = NewFileStream(fsPath, s.settings, s.logger)
-		if s.resumeState != nil {
-			for k, v := range s.resumeState.FileStreamOffset {
-				s.fileStream.SetOffset(k, v)
-			}
-		}
+		s.fileStream = fs.NewFileStream(
+			fs.WithPath(fsPath),
+			fs.WithSettings(s.settings),
+			fs.WithLogger(s.logger),
+			fs.WithHttpClient(NewRetryClient(s.settings.GetApiKey().GetValue(), s.logger)),
+			fs.WithOffsets(s.resumeState.FileStreamOffset),
+		)
 		s.fileStream.Start()
 		s.uploader = uploader.NewUploader(s.ctx, s.logger)
 		s.uploader.Start()
@@ -362,11 +364,11 @@ func (s *Sender) checkAndUpdateResumeState(run *service.RunRecord) error {
 	}
 
 	if s.resumeState.FileStreamOffset == nil {
-		s.resumeState.FileStreamOffset = make(map[chunkFile]int)
+		s.resumeState.FileStreamOffset = make(fs.FileStreamOffsetMap)
 	}
-	s.resumeState.FileStreamOffset[historyChunk] = *bucket.GetHistoryLineCount()
-	s.resumeState.FileStreamOffset[eventsChunk] = *bucket.GetEventsLineCount()
-	s.resumeState.FileStreamOffset[outputChunk] = *bucket.GetLogLineCount()
+	s.resumeState.FileStreamOffset[fs.HistoryChunk] = *bucket.GetHistoryLineCount()
+	s.resumeState.FileStreamOffset[fs.EventsChunk] = *bucket.GetEventsLineCount()
+	s.resumeState.FileStreamOffset[fs.OutputChunk] = *bucket.GetLogLineCount()
 
 	// If we are unable to parse the config, we should fail if resume is set to must
 	// for any other case of resume status, it is fine to ignore it
